pyproject.toml
[project]
name = "fitness-mcp"
version = "0.1.0"
description = "MCP server for memory storage with full-text search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.3",
    "logfire[sqlalchemy]>=4.10.0",
    "psycopg[binary]>=3.2.10",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
